Final Fourth Meal App
Designed the app a little bit and put in the functioning features of
notifications at 9:30pm every night and changing the text on the app
depending on the day. I want to also add images of what the food looks
like on the notifications and/or below the app label.